size_t changes or other not significant changhes

+++ exi/projectExiConnector.h
@@ -17,8 +17,8 @@ extern struct dinEXIDocument dinDocEnc; /* The DIN document. For encoder. */
 extern struct dinEXIDocument dinDocDec; /* The DIN document. For decoder. */
 extern bitstream_t global_streamEnc; /* The byte stream descriptor. */
 extern bitstream_t global_streamDec; /* The byte stream descriptor. */
-extern uint32_t global_streamEncPos; /* The position in the stream. */
-extern uint32_t global_streamDecPos; /* The position in the stream. */
+extern size_t global_streamEncPos; /* The position in the stream. */
+extern size_t global_streamDecPos; /* The position in the stream. */
 extern int g_errn;
 
 #define SESSIONID_LEN 8
